requirements.txt
astra-toolbox
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.59.1
imagecodecs
imageio==2.37.0
importlib_resources==6.5.2
Jinja2
kiwisolver==1.4.7
MarkupSafe
matplotlib==3.9.4
mpmath
networkx
numpy
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML
scipy
six
sympy
tifffile
torch==2.1.0
tqdm==4.67.1
typing_extensions
zipp==3.23.0

torchvision
pytorch-msssim
lpips